+++ sonara-python/src/filters.rs
@@ -7,14 +7,14 @@ use sonara::filters as rs;
 #[pyo3(name = "mel", signature = (*, sr=22050.0, n_fft=2048, n_mels=128, fmin=0.0, fmax=0.0, htk=false, norm="slaney"))]
 pub fn py_mel<'py>(
     py: Python<'py>,
-    sr: f64,
+    sr: f32,
     n_fft: usize,
     n_mels: usize,
-    fmin: f64,
-    fmax: f64,
+    fmin: f32,
+    fmax: f32,
     htk: bool,
     norm: &str,
-) -> Bound<'py, PyArray2<f64>> {
+) -> Bound<'py, PyArray2<f32>> {
     rs::mel(sr, n_fft, n_mels, fmin, fmax, htk, norm).into_pyarray(py)
 }
